+++ CHANGELOG.md
@@ -500,3 +500,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix unlock crash on victory with custom characters (kiooeht)
 * Fix PostDeath hook not triggering on killing Heart (kiooeht)
 * Tag the basegame basic cards with BASIC_STRIKE, BASIC_DEFEND, and FORM (kiooeht)
+* Custom victory cutscenes for custom characters (kiooeht)